+++ clang-tools-extra/clang-tidy/utils/LexerUtils.cpp
@@ -31,6 +31,67 @@ Token getPreviousToken(SourceLocation Location, const SourceManager &SM,
   return Token;
 }
 
+SourceLocation findPreviousTokenStart(SourceLocation Start,
+                                      const SourceManager &SM,
+                                      const LangOptions &LangOpts) {
+  if (Start.isInvalid() || Start.isMacroID())
+    return SourceLocation();
+
+  SourceLocation BeforeStart = Start.getLocWithOffset(-1);
+  if (BeforeStart.isInvalid() || BeforeStart.isMacroID())
+    return SourceLocation();
+
+  return Lexer::GetBeginningOfToken(BeforeStart, SM, LangOpts);
+}
+
+SourceLocation findPreviousTokenKind(SourceLocation Start,
+                                     const SourceManager &SM,
+                                     const LangOptions &LangOpts,
+                                     tok::TokenKind TK) {
+  while (true) {
+    SourceLocation L = findPreviousTokenStart(Start, SM, LangOpts);
+    if (L.isInvalid() || L.isMacroID())
+      return SourceLocation();
+
+    Token T;
+    if (Lexer::getRawToken(L, T, SM, LangOpts, /*IgnoreWhiteSpace=*/true))
+      return SourceLocation();
+
+    if (T.is(TK))
+      return T.getLocation();
+
+    Start = L;
+  }
+}
+
+SourceLocation findNextTerminator(SourceLocation Start, const SourceManager &SM,
+                                  const LangOptions &LangOpts) {
+  return findNextAnyTokenKind(Start, SM, LangOpts, tok::comma, tok::semi);
+}
+
+bool rangeContainsExpansionsOrDirectives(SourceRange Range,
+                                         const SourceManager &SM,
+                                         const LangOptions &LangOpts) {
+  assert(Range.isValid() && "Invalid Range for relexing provided");
+  SourceLocation Loc = Range.getBegin();
+
+  while (Loc < Range.getEnd()) {
+    if (Loc.isMacroID())
+      return true;
+
+    llvm::Optional<Token> Tok = Lexer::findNextToken(Loc, SM, LangOpts);
+
+    if (!Tok)
+      return true;
+
+    if (Tok->is(tok::hash))
+      return true;
+
+    Loc = Lexer::getLocForEndOfToken(Loc, 0, SM, LangOpts).getLocWithOffset(1);
+  }
+
+  return false;
+}
 } // namespace lexer
 } // namespace utils
 } // namespace tidy