+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -136,6 +136,12 @@ Improvements to clang-tidy
   Detects usage of the deprecated member types of ``std::ios_base`` and replaces
   those that have a non-deprecated equivalent.
 
+- New :doc:`readability-isolate-decl
+  <clang-tidy/checks/readability-isolate-decl>` check.
+
+  Detects local variable declarations declaring more than one variable and
+  tries to refactor the code to one statement per declaration.
+
 - New :doc:`readability-magic-numbers
   <clang-tidy/checks/readability-magic-numbers>` check.